Added generic background image support
