+++ app/src/main/java/fi/metatavu/muisti/exhibitionui/pages/components/AbstractComponentFactory.kt
@@ -2,7 +2,10 @@ package fi.metatavu.muisti.exhibitionui.pages.components
 
 import android.annotation.SuppressLint
 import android.content.Context
+import android.content.res.Resources
+import android.graphics.BitmapFactory
 import android.graphics.Color
+import android.graphics.drawable.BitmapDrawable
 import android.net.Uri
 import android.text.Html
 import android.text.Spanned
@@ -21,8 +24,10 @@ import fi.metatavu.muisti.api.client.models.PageLayoutView
 import fi.metatavu.muisti.api.client.models.PageLayoutViewProperty
 import fi.metatavu.muisti.api.client.models.VisitorSession
 import fi.metatavu.muisti.exhibitionui.ExhibitionUIApplication
+import fi.metatavu.muisti.exhibitionui.pages.PageViewVisitorSessionListener
 import fi.metatavu.muisti.exhibitionui.script.ScriptController
 import fi.metatavu.muisti.exhibitionui.session.VisitorSessionContainer
+import fi.metatavu.muisti.exhibitionui.views.PageActivity
 import uk.co.deanwild.flowtextview.FlowTextView
 import java.io.File
 import java.io.FileOutputStream
@@ -43,24 +48,14 @@ abstract class AbstractComponentFactory<T : View> : ComponentFactory<T> {
     private val displayMetrics: DisplayMetrics = context.resources.displayMetrics
 
     /**
-     * Sets view id
+     * Sets up view
      *
+     * @param buildContext build context
      * @param view view
-     * @param pageLayoutView page layout view
      */
-    protected fun setId(view: T, pageLayoutView: PageLayoutView) {
-        setId(view, pageLayoutView.id)
-    }
-
-    /**
-     * Sets view id
-     *
-     * @param view view
-     * @param value value
-     */
-    protected fun setId(view: T, value: String?) {
-        view.tag = value
-        view.transitionName = value
+    protected fun setupView(buildContext: ComponentBuildContext, view: T) {
+        setId(view, buildContext.pageLayoutView)
+        initializeScriptedListener(buildContext, view)
     }
 
     /**
@@ -93,6 +88,7 @@ abstract class AbstractComponentFactory<T : View> : ComponentFactory<T> {
             "layout_toRightOf" -> setLayoutOf(view, property)
             "layout_gravity" -> setLayoutGravity(view, property.value)
             "background" -> setBackground(view, property.value)
+            "backgroundImage" -> setBackgroundImage(buildContext, view, property.value)
             "elevation" -> setElevation(view, property)
             else -> Log.d(ImageViewComponentFactory::javaClass.name, "Property ${property.name} not supported on ${view.javaClass.name} view")
         }
@@ -302,6 +298,66 @@ abstract class AbstractComponentFactory<T : View> : ComponentFactory<T> {
     }
 
     /**
+     * Initializes scripted resource update listeners
+     *
+     * @param buildContext build context
+     * @param view view
+     */
+    private fun initializeScriptedListener(buildContext: ComponentBuildContext, view: T) {
+        buildContext.addVisitorSessionListener(object : PageViewVisitorSessionListener {
+
+            override suspend fun prepareVisitorSessionChange(pageActivity: PageActivity, visitorSession: VisitorSession) {
+                prepareBackgroundImage(visitorSession)
+            }
+
+            override fun performVisitorSessionChange(pageActivity: PageActivity, visitorSession: VisitorSession) {
+                updateBackgroundImage(visitorSession)
+            }
+
+            /**
+             * Prepares background image for scripted resources
+             */
+            private fun prepareBackgroundImage(visitorSession: VisitorSession) {
+                val url = getUrl(getScriptedResource(buildContext,  visitorSession, "backgroundImage", false))
+                if (url != null) {
+                    getResourceOfflineFile(buildContext, url)
+                }
+            }
+
+            /**
+             * Updates background image for scripted resources
+             */
+            private fun updateBackgroundImage(visitorSession: VisitorSession) {
+                val url = getUrl(getScriptedResource(buildContext, visitorSession,"backgroundImage", false))
+                if (url != null) {
+                    setBackgroundImage(buildContext, view, url)
+                }
+            }
+        })
+    }
+
+    /**
+     * Sets view id
+     *
+     * @param view view
+     * @param pageLayoutView page layout view
+     */
+    private fun setId(view: T, pageLayoutView: PageLayoutView) {
+        setId(view, pageLayoutView.id)
+    }
+
+    /**
+     * Sets view id
+     *
+     * @param view view
+     * @param value value
+     */
+    private fun setId(view: T, value: String?) {
+        view.tag = value
+        view.transitionName = value
+    }
+
+    /**
      * Evaluates scripted resource value
      *
      * @param resource scripted resource
@@ -333,6 +389,49 @@ abstract class AbstractComponentFactory<T : View> : ComponentFactory<T> {
         val color = getColor(value)
         color ?: return
         view.setBackgroundColor(color)
+    }
+
+    /**
+     * Sets background image.
+     *
+     * On scripted resources this method is does not do anything because scripted resources are
+     * handled by visitor session change events
+     *
+     * @param buildContext build context
+     * @param view view component
+     * @param value value
+     */
+    private fun setBackgroundImage(buildContext: ComponentBuildContext, view: View, value: String?) {
+        val resource = getResource(buildContext, value)
+        val scripted = resource?.scripted ?: false
+
+        if (scripted) {
+            return
+        }
+
+        val resourceData = resource?.data
+
+        val url = getUrl(resourceData ?: value)
+        url ?: return
+
+        setBackgroundImage(buildContext = buildContext, view = view, url = url)
+    }
+
+    /**
+     * Sets background image
+     *
+     * @param buildContext build context
+     * @param view view component
+     * @param url url
+     */
+    private fun setBackgroundImage(buildContext: ComponentBuildContext, view: View, url: URL?) {
+        val offlineFile = getResourceOfflineFile(buildContext = buildContext, url = url)
+        offlineFile ?: return
+
+        val bitmap = BitmapFactory.decodeFile(offlineFile.absolutePath)
+        if (bitmap != null) {
+            view.background = BitmapDrawable(Resources.getSystem(), bitmap)
+        }
     }
 
     /**
